# Optimize router