Hi, 

This PR attempts to optimize the router by pre-compiling pattern.

### What did I do?

* Call `getPattern` at `insert` phase. And cache them.
* Do loop only for patterns.
* Skip regexp capturing. I think the pattern for path meaning "accept current component or not".

### Breaking change

I think `/users/:id{[0-9]+}` used to match `/users/abc123xyz`, but with this change, it will match `/^[0-9]+$/` strictly.

### Benchmark

With this PR, I got the following results.

```
hono x 413,513 ops/sec ±4.13% (71 runs sampled)
itty-router x 70,685 ops/sec ±3.19% (86 runs sampled)
sunder x 124,893 ops/sec ±3.05% (85 runs sampled)
worktop x 76,740 ops/sec ±4.47% (77 runs sampled)
Fastest is hono
```

### More

I have another idea about improving the performance of the router, and I hope to make another PR about it in the few days.