@yusukebe 2022-02-18
Hi @usualoma 

This is cool! Benchmark results on my PC are here:

Before:
```
hono x 734,434 ops/sec ±3.99% (69 runs sampled)
itty-router x 161,362 ops/sec ±3.66% (90 runs sampled)
sunder x 301,725 ops/sec ±5.08% (75 runs sampled)
worktop x 212,848 ops/sec ±4.18% (85 runs sampled)
Fastest is hono
```

After:
```
hono x 779,197 ops/sec ±6.55% (78 runs sampled)
itty-router x 161,813 ops/sec ±3.87% (87 runs sampled)
sunder x 334,096 ops/sec ±1.33% (93 runs sampled)
worktop x 212,661 ops/sec ±4.40% (81 runs sampled)
Fastest is hono
```

Really ultrafast! Thank you, I'll merge it.